TwitterBot retweet can now be called through a rake task.
