+++ lib/TwitterBot.rb
@@ -1,6 +1,17 @@
 require "twitter"
+require "dotenv-rails"
 
-class TwitterBot < ApplicationController
+class TwitterBot
+
+  def twitter_client
+    Twitter::REST::Client.new do |config|
+      config.consumer_key = ENV['CONSUMER_KEY']
+      config.consumer_secret = ENV['CONSUMER_SECRET']
+      config.access_token = ENV['ACCESS_TOKEN']
+      config.access_token_secret = ENV['ACCESS_SECRET']
+    end
+  end
+
   def retweet
     # Snag the search results
     search
